misc + fix(actions, closes #18): (#19)
* misc: restructuring src directory and separating entrypoint with backend code

* fix(actions, closes #18): action yml file now references the correct directory

* fix(actions, closes #18): switching back to original yml path, since it was correct
