+++ src/backend.go
@@ -1,14 +1,11 @@
-// Package main provides the backend infrastructure (route handling + database) 
+// Package backend provides the backend infrastructure (route handling + database) 
 // for the Bricked-Up website.
-package main
+package backend
 
 import (
     "fmt"
-    "log"
     "net/http"
 )
-
-const PORT = ":3100" // Server port number.
 
 // Endpoints maps URL paths to their corresponding handler functions.
 var endpoints = map[string]http.HandlerFunc{
@@ -49,21 +46,10 @@ func verifyHandler(w http.ResponseWriter, r *http.Request) {
 
 // MainHandler checks if the request URL matches a known endpoint.
 // If it does, the corresponding handler is called; otherwise, it returns a 404 error.
-func mainHandler(w http.ResponseWriter, r *http.Request) {
+func MainHandler(w http.ResponseWriter, r *http.Request) {
     if handler, ok := endpoints[r.URL.Path]; ok {
         handler(w, r)
         return
     }
     http.NotFound(w, r)
-}
-
-// Main sets up the server and starts listening on the defined PORT.
-// It registers MainHandler to process all incoming HTTP requests.
-func main() {
-    http.HandleFunc("/", mainHandler)
-
-    log.Printf("Listening on localhost%s", PORT)
-    if err := http.ListenAndServe(PORT, nil); err != nil {
-        log.Fatal(err)
-    }
 }
